+++ neural/layer.h
@@ -9,18 +9,18 @@ class layer
 {
 private:
     /* data */
-    std::vector<std::vector<double>> weight;
-    std::vector<double> bias;
+    std::vector<std::vector<long double>> weight;
+    std::vector<long double> bias;
     //pointer to input data
     
     
 public:
     layer(int weights, int neurons);
     ~layer();
-    double randomDouble();
+    long double randomDouble();
     void testCout();
-    double getWeights(int row, int columns);
-    double getBias(int i);
+    long double getWeights(int row, int columns);
+    long double getBias(int i);
     int getSizeRows();
     int getSizeColumns(int i);
